Flow for getLanguage, getPrefixedPath rename
